Brand Page v1.0 + Logo Assets
The Brand page has been created to have a quick access to all of our assets. In it you can find links for the downloads of:
- Brand Guidelines
- Logo packs on all it's versions

The brand folder contains an AI CS5 compatible file that has every version combine in it's color version.
An animated version of the logo was added as a creative way to present it.

Credits in this release:
@ostettd#3098 Correcting and perfecting of the logo